Move error.rs to root of crate
