+++ rust/datafusion/src/lib.rs
@@ -25,6 +25,7 @@ extern crate serde_json;
 extern crate sqlparser;
 
 pub mod datasource;
+pub mod error;
 pub mod execution;
 pub mod logicalplan;
 pub mod optimizer;
